providers/langchain/pyproject.toml
[project]
name = "lab-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "langchain-openai>=0.3.24",
    "langchain[ollama,openai]>=0.3.26",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
